admaren_test/admaren/requirement.txt
Django==3.0.6
djangorestframework==3.11.0
mysqlclient==1.4.6
djangorestframework-simplejwt